Gephi.

Make your workflow with network data easier. The gephi package does not interface with the open source network visualisiation software 'gephi' but only writes csv files in a format that can be easily read by 'gephi'.

Graphs, networks, edges and nodes

Network science is a broad field with many practicitioners who, unfortunately don't use the same languge to describe the same things. A graph is a network, not to be confused with graphics, or a graph like a scatterplot or line chart. A network (known as graph) consists of points (known as vertices or nodes ) and connections between them (known as vertices, links or edges). The gephi network visualisation tool uses edges and nodes as description, igraph talks about graphs G(), vertices V() and edges E(), tidygraph talks about edges and nodes. We don't care as long as we can work with it.